introduced game module
The game module now contains all the game logic (moved most of the
stuff from game_state) and implemented it as a state-machine. Also
changed the way how moves are represented. Splitted the two parts of a
move into their own actions. This makes the logic more easier to
understand as we have less edge cases to check for.